+++ app/layout.tsx
@@ -1,0 +1,1 @@
+export default function RootLayout({children}) { return <html><body>{children}</body></html> }
